+++ tests/tools/apps/c/msgr-send.c
@@ -32,7 +32,7 @@
 
 typedef struct {
     Addresses_t targets;
-    uint64_t msg_count;
+    unsigned long long msg_count;
     uint32_t msg_size;  // of body
     uint32_t send_batch;
     int   outgoing_window;
@@ -86,7 +86,7 @@ static void parse_options( int argc, char **argv, Options_t *opts )
         switch(c) {
         case 'a': addresses_merge( &opts->targets, optarg ); break;
         case 'c':
-            if (sscanf( optarg, "%lu", &opts->msg_count ) != 1) {
+            if (sscanf( optarg, "%llu", &opts->msg_count ) != 1) {
                 fprintf(stderr, "Option -%c requires an integer argument.\n", optopt);
                 usage(1);
             }
@@ -175,8 +175,8 @@ int main(int argc, char** argv)
 {
     Options_t opts;
     Statistics_t stats;
-    uint64_t sent = 0;
-    uint64_t received = 0;
+    unsigned long long sent = 0;
+    unsigned long long received = 0;
     int target_index = 0;
     int rc;
 
@@ -254,7 +254,7 @@ int main(int argc, char** argv)
         check_messenger(messenger);
     }
 
-    LOG("Messages received=%lu sent=%lu\n", received, sent);
+    LOG("Messages received=%llu sent=%llu\n", received, sent);
 
     if (get_replies) {
         // wait for the last of the replies
@@ -264,7 +264,7 @@ int main(int argc, char** argv)
             check( count > 0 || (opts.timeout == 0),
                    "Error: timed out waiting for reply messages\n");
             received += count;
-            LOG("Messages received=%lu sent=%lu\n", received, sent);
+            LOG("Messages received=%llu sent=%llu\n", received, sent);
         }
     } else if (pn_messenger_outgoing(messenger) > 0) {
         LOG("Calling pn_messenger_send()\n");
